revert init_logging move
